+++ src/Datasets/Generators/HalfMoon.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rubix\ML\Other\Generators;
+namespace Rubix\ML\Datasets\Generators;
 
 use Rubix\ML\Datasets\Dataset;
 use Rubix\ML\Datasets\Unlabeled;
@@ -10,25 +10,25 @@ use MathPHP\Probability\Distribution\Continuous\Uniform;
 use InvalidArgumentException;
 
 /**
- * Circle
+ * Half Moon
  *
- * Create a circle made of sample data points in 2 dimensions.
+ * Generate a dataset consisting of 2-d samples that form a half moon shape.
  *
  * @category    Machine Learning
  * @package     Rubix/ML
  * @author      Andrew DalPino
  */
-class Circle implements Generator
+class HalfMoon implements Generator
 {
     /**
-     * The x and y coordinates of the center of the circle.
+     * The x and y coordinates of the center of the half moon.
      *
      * @var array
      */
     protected $center;
 
     /**
-     * The scaling factor of the circle.
+     * The scaling factor of the moon.
      *
      * @var float
      */
@@ -51,20 +51,27 @@ class Circle implements Generator
     /**
      * @param  array  $center
      * @param  float  $scale
+     * @param  float  $rotate
      * @param  float  $noise
      * @throws \InvalidArgumentException
      * @return void
      */
-    public function __construct(array $center = [0, 0], float $scale = 1.0, float $noise = 0.1)
+    public function __construct(array $center = [0, 0], float $scale = 1.0, float $rotate = 90.0,
+                                float $noise = 0.1)
     {
         if (count($center) !== 2) {
             throw new InvalidArgumentException('This generator only works in 2'
                 . ' dimensions.');
         }
 
-        if ($scale < 0.) {
+        if ($scale < 0) {
             throw new InvalidArgumentException('Scaling factor must be greater'
                 . ' than 0.');
+        }
+
+        if ($rotate < 0 or $rotate > 360) {
+            throw new InvalidArgumentException('Rotation must be between 0 and'
+                . ' 360 degrees.');
         }
 
         if ($noise <= 0.or $noise > 1.) {
@@ -74,7 +81,7 @@ class Circle implements Generator
 
         $this->center = $center;
         $this->scale = $scale;
-        $this->uniform = new Uniform(0, 2. * M_PI);
+        $this->uniform = new Uniform(deg2rad($rotate), deg2rad($rotate + 180));
         $this->gaussian = new Normal(0, $scale * $noise);
     }
 
@@ -111,4 +118,5 @@ class Circle implements Generator
 
         return new Unlabeled($samples, false);
     }
+
 }