+++ src/Persisters/Filesystem.php
@@ -1,11 +1,22 @@
 <?php
 
-namespace Rubix\ML\Other\Persisters;
+namespace Rubix\ML\Persisters;
 
 use Rubix\ML\Persistable;
 use InvalidArgumentException;
 use RuntimeException;
 
+/**
+ * Filesystem
+ *
+ * Filesystems are local or remote storage drives that are organized by files
+ * and folders. The filesystem persister serializes models to a file at a
+ * user-specified path.
+ *
+ * @category    Machine Learning
+ * @package     Rubix/ML
+ * @author      Andrew DalPino
+ */
 class Filesystem implements Persister
 {
     /**
@@ -28,7 +39,7 @@ class Filesystem implements Persister
      * @throws \InvalidArgumentException
      * @return void
      */
-    public function __construct(string $path, bool $overwrite = false)
+    public function __construct(string $path, bool $overwrite = true)
     {
         if (!is_writable(dirname($path))) {
             throw new InvalidArgumentException('Folder does not exist or is not'
@@ -37,6 +48,16 @@ class Filesystem implements Persister
 
         $this->path = $path;
         $this->overwrite = $overwrite;
+    }
+
+    /**
+     * Return the size of the file on the filesystem in bytes.
+     *
+     * @return int|null
+     */
+    public function size() : ?int
+    {
+        return filesize($this->path) ?: null;
     }
 
     /**
@@ -65,15 +86,12 @@ class Filesystem implements Persister
     {
         if (!file_exists($this->path)) {
             throw new RuntimeException('File ' . basename($this->path)
-                . ' cannot be opened. Check the path.');
+                . ' does not exist. Check the path.');
         }
 
-        if (!is_readable($this->path)) {
-            throw new RuntimeException('File ' . basename($this->path)
-                . ' cannot be opened. Check the file permissions.');
-        }
+        $data = file_get_contents($this->path) ?: '';
 
-        $persistable = unserialize(file_get_contents($this->path) ?: '');
+        $persistable = unserialize($data);
 
         if (!$persistable instanceof Persistable) {
             throw new RuntimeException('Object cannot be reconstituted.');
@@ -93,14 +111,26 @@ class Filesystem implements Persister
     public function save(Persistable $persistable) : void
     {
         if ($this->overwrite === false and file_exists($this->path)) {
-            throw new RuntimeException('Attempting to overwrite an existing'
-                . ' model.');
+            throw new RuntimeException('Cannot overwrite existing file.');
         }
 
-        $success = file_put_contents($this->path, serialize($persistable), LOCK_EX);
+        $data = serialize($persistable);
+
+        $success = file_put_contents($this->path, $data, LOCK_EX);
 
         if (!$success) {
-            throw new RuntimeException('Failed to serialize object to a file.');
+            throw new RuntimeException('Failed to serialize object to'
+                . ' filesystem.');
         }
+    }
+
+    /**
+     * Remove the file from the filesystem.
+     *
+     * @return void
+     */
+    public function delete() : void
+    {
+        unlink($this->path);
     }
 }